pyproject.toml
[project]
name = "speechsnap"
version = "0.1.0"
description = "Push-to-Talk speech-to-text application for Linux (Wayland)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["speech-to-text", "whisper", "ptt", "linux", "wayland"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "pydantic-settings",
    "python-dotenv",
    "evdev>=1.9.2",
    "sounddevice>=0.5.3",
    "numpy>=2.4.1",
    "soundfile>=0.13.1",
    "faster-whisper>=1.0.0",
    "torch>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
